+++ AlchemistCode/Potions/Solvent.cs
@@ -1,3 +1,4 @@
+using System.Linq;
 using Alchemist.AlchemistCode.Compat;
 using MegaCrit.Sts2.Core.Commands;
 using MegaCrit.Sts2.Core.Entities.Cards;
@@ -12,22 +13,23 @@ namespace Alchemist.AlchemistCode.Potions;
 
 public class Solvent : AlchemistPotion, IBrewOnly
 {
-    private const int Weak = 3;
 
     public override PotionRarity Rarity => PotionRarity.Event;
     public override PotionUsage Usage => PotionUsage.CombatOnly;
-    public override TargetType TargetType => TargetType.AnyEnemy;
+    public override TargetType TargetType => TargetType.AllEnemies;
 
     public override IEnumerable<IHoverTip> ExtraHoverTips =>
-        new[] { HoverTipFactory.FromPower<ArtifactPower>(), HoverTipFactory.FromPower<WeakPower>() };
+        new[] { HoverTipFactory.FromPower<ArtifactPower>() };
 
     protected override async Task OnUse(PlayerChoiceContext choiceContext, Creature? target)
     {
-        PotionModel.AssertValidForTargetedPotion(target);
-        if (target!.Block > 0)
-            await GameCompat.LoseBlock(choiceContext, target, target.Block, Owner.Creature);
-        if (target.HasPower<ArtifactPower>())
-            await PowerCmd.Remove<ArtifactPower>(target);
-        await PowerCmd.Apply<WeakPower>(choiceContext, target, Weak, Owner.Creature, null);
+        if (Owner.Creature.CombatState is not { } combat) return;
+        foreach (var enemy in combat.HittableEnemies.ToList())
+        {
+            if (enemy.Block > 0)
+                await GameCompat.LoseBlock(choiceContext, enemy, enemy.Block, Owner.Creature);
+            if (enemy.HasPower<ArtifactPower>())
+                await PowerCmd.Remove<ArtifactPower>(enemy);
+        }
     }
 }